refactor(team): split launcher_web.go and broker_lifecycle.go (C8)
Pure file split per PLAN.md §C8. No new types, no behaviour changes —
just moves functions out of launcher.go into themed sibling files so
the orchestrator file stops being the catch-all.

launcher_web.go (304 lines, new) — web-mode entry points:
  PreflightWeb, LaunchWeb, maybeOfferNex, waitForWebReady, stdinIsTTY,
  openBrowser. Lets web-only imports (`net`, `golang.org/x/term`,
  `internal/setup`, `internal/nex`) sit in one file instead of bloating
  launcher.go's import list.

broker_lifecycle.go (148 lines, new) — broker housekeeping:
  killStaleBroker, ResetBrokerState, ClearPersistedBrokerState,
  officePIDFilePath, writeOfficePIDFile, clearOfficePIDFile,
  killPersistedOfficeProcess, resetBrokerState, brokerBaseURL, and
  Launcher.BrokerBaseURL. Groups the office-PID file lifecycle, the
  in-memory + persisted broker-state reset endpoints, and the
  port-killer helpers in one place.

launcher.go: 3431 → 3034 lines (-397 in this PR; cumulative since main:
4998 → 3034 = -1964 lines, -39.3%).

Imports trimmed from launcher.go: io, net, errors, strconv,
golang.org/x/term, internal/nex, internal/setup. Package coverage and
per-file gates unchanged — these are existing helpers and are still
exercised by the same tests.

Stacked on top of refactor/launcher-c7-headless-pool (C7).
